+++ source/putting_it_together_ch02.cpp
@@ -14,11 +14,11 @@ void CH02_putting_it_together()
      auto ca = c_allocator_get();
 
     Point start = point(0, 1, 0);
-    Vector velocity = vector_scale(vector_normalized(vector(1, 1.8, 0)), 11.25);
+    Vector velocity = vector_scale(vector_normalized(vector(1.f, 1.8f, 0.f)), 11.25f);
     Projectile p = projectile(start, velocity);
 
-    Vector gravity = vector(0, -0.1, 0);
-    Vector wind = vector(-0.01, 0, 0);
+    Vector gravity = vector(0.f, -0.1f, 0.f);
+    Vector wind = vector(-0.01f, 0.f, 0.f);
     Environment e = environment(gravity, wind);
 
     Canvas c = canvas(900, 550, ca);
@@ -29,15 +29,15 @@ void CH02_putting_it_together()
     while (p.position.y > 0) {
 
         Point pixel_pos = point(p.position.x, c.height - p.position.y, p.position.z);
-        pixel_pos.x = clamp(0, c.width - 1, pixel_pos.x);
-        pixel_pos.y = clamp(0, c.height - 1, pixel_pos.y);
-        canvas_set_pixel(&c, pixel_pos.x, pixel_pos.y, p_color);
+        pixel_pos.x = (float)clamp(0, c.width - 1, pixel_pos.x);
+        pixel_pos.y = (float)clamp(0, c.height - 1, pixel_pos.y);
+        canvas_set_pixel(&c, (int)pixel_pos.x, (int)pixel_pos.y, p_color);
         p = tick(e, p);
     }
     Point pixel_pos = point(p.position.x, c.height - p.position.y, p.position.z);
         pixel_pos.x = clamp(0, c.width - 1, pixel_pos.x);
         pixel_pos.y = clamp(0, c.height - 1, pixel_pos.y);
-    canvas_set_pixel(&c, pixel_pos.x, pixel_pos.y, p_color);
+    canvas_set_pixel(&c, (int)pixel_pos.x, (int)pixel_pos.y, p_color);
 
     String ppm_str = canvas_to_ppm(c, ca);
     printf("%s\n", ppm_str.data);
